requirements.txt
Flask==2.2.3
requests==2.28.2
python-dotenv==1.0.0
pyngrok==6.0.0
openai==1.3.5
firebase-admin==6.2.0
geopy==2.3.0